Move admin config handlers to handler package
