+++ server/router.go
@@ -31,6 +31,7 @@ func InitRoute() error {
 	routes.InitCommentRoute(ginServer)
 	routes.InitAiRouter(ginServer)
 	routes.InitAdminRoute(ginServer)
+	routes.InitSystemRoute(ginServer)
 	routes.InitMiscRoute(ginServer)
 
 	return nil
